Update Inventory for each order + Disabled Add to card and reorder features for unauthorized stores

+++ app/src/main/java/com/example/test/Component/OrderedProduct.kt
@@ -319,7 +319,7 @@ fun OrderedProduct(data: HashMap<String, Any?>, onLeftSwipe: SwipeAction , index
                                     modifier = Modifier
                                         .clickable(
                                             enabled = quantity < data["inventory"]
-                                                .toString()
+                                                .toString().toFloat()
                                                 .toInt(),
                                             onClick = {
